fix(participants): stop call to api if user do not have permission to add participants to a study
