+++ src/components/studyDetails/ParticipantComponent.tsx
@@ -51,9 +51,9 @@ const ParicipantComponent: React.FC<ParicipantComponentProps> = ({ study, setStu
     const [selectedParticipant, setSelectedParticipant] = useState<ParticipantObj | undefined>();
     const [text, setText] = useState<string>('Search or add by e-mail');
     const [role, setRole] = useState<string>('');
-    const { loading, setLoading } = useFetch(api.getStudyRoles, setRoles);
+    const { loading, setLoading } = useFetch(api.getStudyRoles, setRoles, 'studyRoles');
 
-    useEffect(() => {}, [participantNotSelected]);
+    useEffect(() => {}, [participantNotSelected, study.permissions.addRemoveParticipant]);
 
     const removeParticipant = (participant: any) => {
         let participantList: any = [...study.participants];
@@ -118,6 +118,28 @@ const ParicipantComponent: React.FC<ParicipantComponentProps> = ({ study, setStu
         return tempRoles;
     };
 
+    const getParticipants = (inputValue: string, callback: any): void => {
+        api.getParticipantList(inputValue || 'a').then((result: any) => {
+            if (!result.Message) {
+                let temp = result.map((user) => {
+                    return {
+                        label: `${user.fullName} (${user.emailAddress})`,
+                        value: user.objectId,
+                        emailAddress: user.emailAddress,
+                        source: user.source,
+                        objectId: user.objectId,
+                        name: user.fullName,
+                        databaseId: user.databaseId
+                    };
+                });
+                callback(temp);
+            } else {
+                console.log('err');
+                //notify.show('danger', '500');
+            }
+        });
+    };
+
     const selectParticipant = (row: any) => {
         setText(row.label);
         setParticipantNotSelected(false);
@@ -172,6 +194,9 @@ const ParicipantComponent: React.FC<ParicipantComponentProps> = ({ study, setStu
                             selectedOption={{ value: 'Search..', label: text }}
                             onInputChange={handleInputChange}
                             disabled={study.permissions && !study.permissions.addRemoveParticipant}
+                            loadOptions={
+                                study.permissions && study.permissions.addRemoveParticipant ? getParticipants : null
+                            }
                         />
                     </div>
                 </Tooltip>
